feat!: adds button callback example, moves awtrix config into broker package
